+++ R/mod_uploadPackage.R
@@ -73,7 +73,7 @@ uploadPackageUI <- function(id) {
 #' 
 #' @importFrom riskmetric pkg_ref
 #' @importFrom rintrojs introjs
-#' @importFrom utils read.csv available.packages
+#' @importFrom utils read.csv available.packages download.file
 #' @importFrom rvest read_html html_nodes html_text
 #' @keywords internal
 #' 
@@ -308,7 +308,7 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
         good_urls <- purrr::map_lgl(url_lst, 
                                     ~ try(curlGetHeaders(.x, verify = FALSE), silent = TRUE) %>%
                                       {class(.) != "try-error" && attr(., "status") != 404})
-        
+
         if (!all(good_urls)) {
           checking_urls$url_lst <- url_lst[!good_urls]
           showModal(modalDialog(
@@ -343,10 +343,11 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
         message = "Uploading Packages to DB:", {
           
           for (i in 1:np) {
-            
+
             user_ver <- uploaded_packages$version[i]
             incProgress(1, detail = glue::glue("{uploaded_packages$package[i]} {user_ver}"))
             
+
             if (grepl("^[[:alpha:]][[:alnum:].]*[[:alnum:]]$", uploaded_packages$package[i])) {
               # run pkg_ref() to get pkg version and source info
               if (!isTRUE(getOption("shiny.testmode")))
@@ -361,39 +362,30 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
 
             if (ref$source %in% c("pkg_missing", "name_bad")) {
               incProgress(1, detail = 'Package {uploaded_packages$package[i]} not found')
-              
+
               # Suggest alternative spellings using utils::adist() function
               v <- utils::adist(uploaded_packages$package[i], cran_pkgs(), ignore.case = FALSE)
               rlang::inform(paste("Package name",uploaded_packages$package[i],"was not found."))
-              
+
               suggested_nms <- paste("Suggested package name(s):",paste(head(cran_pkgs()[which(v == min(v))], 10),collapse = ", "))
               rlang::inform(suggested_nms)
-              
+
               uploaded_packages$status[i] <- HTML(paste0('<a href="#" title="', suggested_nms, '">not found</a>'))
-              
+
               if (ref$source == "pkg_missing")
                 loggit::loggit('WARN',
                                glue::glue('Package {ref$name} was flagged by riskmetric as {ref$source}.'))
               else
                 loggit::loggit('WARN',
                                glue::glue("Riskmetric can't interpret '{ref$name}' as a package reference."))
-              
+
               next
             }
-            
+
             ref_ver <- as.character(ref$version)
             
-            # The message about different versions is only given if the package 
-            # name is uploaded from CSV file and not if it is selected from drop-down.
-            if(isTruthy(input$load_cran)) {
-              ver_msg <- ref_ver
-            } else {
-              if(user_ver == ref_ver) ver_msg <- ref_ver
-              else ver_msg <- glue::glue("{ref_ver}, not '{user_ver}'")
-            }
-            
-            as.character(ref$version)
-            deets <- glue::glue("{uploaded_packages$package[i]} {ver_msg}")
+            deets <- glue::glue("{uploaded_packages$package[i]} {ref_ver}")
+            uploaded_packages$version[i] <- ref_ver
             
             # Save version.
             incProgress(1, detail = deets)
@@ -410,8 +402,9 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
             if(!found) {
               # Get and upload pkg general info to db.
               incProgress(1, detail = deets)
+
               if (!isTRUE(getOption("shiny.testmode"))) {
-                dwn_ld <- download.file(ref$tarball_url, file.path("tarballs", basename(ref$tarball_url)), 
+                dwn_ld <- utils::download.file(ref$tarball_url, file.path("tarballs", basename(ref$tarball_url)), 
                                         quiet = TRUE, mode = "wb")
                 if (dwn_ld != 0) {
                   loggit::loggit("INFO", glue::glue("Unable to download the source files for {uploaded_packages$package[i]} from '{ref$tarball_url}'."))
@@ -420,9 +413,11 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
               insert_pkg_info_to_db(uploaded_packages$package[i], ref_ver)
               # Get and upload maintenance metrics to db.
               incProgress(1, detail = deets)
+              
               insert_riskmetric_to_db(uploaded_packages$package[i])
               # Get and upload community metrics to db.
               incProgress(1, detail = deets)
+              
               insert_community_metrics_to_db(uploaded_packages$package[i])
               uploaded_packages$score[i] <- get_pkg_info(uploaded_packages$package[i])$score
               if (!rlang::is_empty(auto_list())) {
@@ -435,9 +430,10 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
           incProgress(1, detail = "   **Completed Pkg Uploads**")
           Sys.sleep(0.25)
           
-        }) #withProgress
+      }) #withProgress
       
       uploaded_pkgs(uploaded_packages)
+      
     })
     
     # Download the sample dataset.
@@ -454,6 +450,7 @@ uploadPackageServer <- function(id, user, auto_list, credentials, trigger_events
     output$upload_summary_text <- renderText({
       req(uploaded_pkgs)
       req(nrow(uploaded_pkgs()) > 0)
+
       # modify the message if we are removing packages
       if(isTruthy(sum(uploaded_pkgs()$status == 'removed') >0)) {
         loggit::loggit("INFO",